+++ src/components/Controllers/Table/Panels/Source.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { useTheme } from '@mui/material/styles'
 import Box from '@mui/material/Box'
-import TextEditor from '../../../Editors/Text'
+import TextEditor from '../../../Parts/Text'
 import { useStore } from '../store'
 
 export default function SourcePanel() {
