+++ src/components/Controllers/Text/store.ts
@@ -8,7 +8,7 @@ import { createStore } from 'zustand/vanilla'
 import { createSelector } from 'reselect'
 import { assert } from 'ts-essentials'
 import { Client } from '../../../client'
-import { ITextEditor } from '../../Editors/Text'
+import { ITextEditor } from '../../Parts/Text'
 import { TextProps } from './index'
 import * as helpers from './helpers'
 import * as types from '../../../types'
