+++ src/components/Parts/Text/index.tsx
